refactor: separate the responsibilities of cognitive functions to isolated files
