+++ src/lib/aws-cognito/helpers/index.tsx
@@ -1,0 +1,4 @@
+export { signUp } from './signUp';
+export { authenticate } from './authenticate';
+export { confirmAccount } from './confirmAccount';
+export { resendConfirmationCode } from './resendConfirmationCode';
